[18110] Refactored type of NumberOfItems field in ChunkEntity into short.
git-svn-id: https://svn.dbc.dk/repos/dataio/trunk@80302 b87a6194-a221-0410-ad6f-3b8ec35bba13

+++ new-job-store-service/ejb/src/main/java/dk/dbc/dataio/jobstore/service/entity/ChunkEntity.java
@@ -27,7 +27,7 @@ public class ChunkEntity {
     private String dataFileId;
 
     @Column(nullable = false)
-    private int numberOfItems;
+    private short numberOfItems;
 
     @Column(insertable = false, updatable = false)
     private Timestamp timeOfCreation;
@@ -61,11 +61,11 @@ public class ChunkEntity {
         this.dataFileId = dataFileId;
     }
 
-    public int getNumberOfItems() {
+    public short getNumberOfItems() {
         return numberOfItems;
     }
 
-    public void setNumberOfItems(int numberOfItems) {
+    public void setNumberOfItems(short numberOfItems) {
         this.numberOfItems = numberOfItems;
     }
